355-rename-lagom-dev-mode-components: rename DevModeServiceDiscovery to LagomDevModeServiceDiscovery and fix import for com.lightbend.lagom.devmode.internal.registry.serviceDnsRecords
