+++ dev/service-registry/devmode-scaladsl/src/main/scala/com/lightbend/lagom/scaladsl/devmode/LagomDevModeComponents.scala
@@ -11,7 +11,7 @@ import akka.actor.CoordinatedShutdown
 import akka.discovery.ServiceDiscovery
 import akka.stream.Materializer
 import com.lightbend.lagom.devmode.internal.scaladsl.registry.ServiceRegistry
-import com.lightbend.lagom.devmode.internal.registry.DevModeServiceDiscovery
+import com.lightbend.lagom.devmode.internal.registry.LagomDevModeServiceDiscovery
 import com.lightbend.lagom.devmode.internal.registry.ServiceRegistryClient
 import com.lightbend.lagom.internal.scaladsl.client.ScaladslServiceClient
 import com.lightbend.lagom.internal.scaladsl.client.ScaladslServiceResolver
@@ -112,8 +112,8 @@ trait LagomDevModeServiceLocatorComponents extends CircuitBreakerComponents {
   private lazy val serviceRegistryClient: ServiceRegistryClient =
     new ScalaServiceRegistryClient(serviceRegistry)(executionContext)
 
-  private lazy val devModeSimpleServiceDiscovery: DevModeServiceDiscovery =
-    DevModeServiceDiscovery(actorSystem)
+  private lazy val devModeSimpleServiceDiscovery: LagomDevModeServiceDiscovery =
+    LagomDevModeServiceDiscovery(actorSystem)
   // This needs to be done eagerly to ensure it initializes for Akka libraries
   // that use service discovery without dependency injection.
   devModeSimpleServiceDiscovery.setServiceRegistryClient(serviceRegistryClient)
